requirements.txt
aiosmtplib==1.1.7
anyio==3.3.4
asgiref==3.6.0
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.5
certifi==2023.11.17
cffi==1.15.0
charset-normalizer==3.3.2
click==8.0.3
cryptography==35.0.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.3.0
fastadmin==0.1.40
fastapi==0.70.0
fastapi-mail==1.1.4
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
isort==5.13.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.31.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0

httpx
apscheduler
